+++ components-ui/app/src/tol-ui/src/table/cellRenderers/CellEditableDatetime.tsx
@@ -1,0 +1,98 @@
+/*
+SPDX-FileCopyrightText: 2026 Genome Research Ltd.
+
+SPDX-License-Identifier: MIT
+*/
+
+import { useState } from "react";
+import { DatePicker } from "rsuite";
+import {
+  CellEditableControls,
+  MESSAGE_TYPE,
+  PCellEditableInput,
+  PopUpMessage,
+  VALUE_CANNOT_BE_BLANK_MESSAGE,
+} from "../..";
+
+
+/**
+ * Date picker used for inline value updates.
+ */
+export function CellEditableDatetime(props: PCellEditableInput) {
+  const {
+    field,
+    value,
+    setValue,
+    dataObject,
+    dataSource,
+    onSaveSuccess,
+    onSaveError,
+    floatingControls,
+    onCancel,
+    loading,
+    setLoading,
+  } = props;
+
+  const [datePickerOpen, setDatePickerOpen] = useState(true);
+
+  const onChange = (newValue: Date) => {
+    setValue(newValue);
+  }
+
+  const onSave = () => {
+    // prevent saving blank values
+    if (typeof value === "string" && value.trim() === "") {
+      PopUpMessage({
+        type: MESSAGE_TYPE.ERROR,
+        message: VALUE_CANNOT_BE_BLANK_MESSAGE,
+      });
+      return;
+    }
+
+    if (!dataObject) return;
+    setLoading(true);
+    dataSource
+      ?.upsert({
+        objectType: dataObject?.objectType,
+        payload: [
+          {
+            type: dataObject?.objectType,
+            id: dataObject?.id,
+            attributes: {
+              [field]: value,
+            },
+          },
+        ],
+      })
+      .then(() => onSaveSuccess?.())
+      .catch(() => onSaveError?.())
+      .finally(() => setLoading(false));
+  }
+
+  return (
+    <>
+      <DatePicker
+        value={new Date(value)}
+        onChange={(date) => {
+          if (!date) return;
+          onChange(date);
+          setDatePickerOpen(false);
+        }}
+        cleanable={false}
+        preventOverflow
+        oneTap
+        block
+        open={datePickerOpen}
+        onOpen={() => setDatePickerOpen(true)}
+        onClose={() => setDatePickerOpen(false)}
+        editable={false}
+      />
+      <CellEditableControls
+        floatingControls={floatingControls}
+        loading={loading}
+        onCancel={onCancel}
+        onSave={onSave}
+      />
+    </>
+  );
+}